Test for empty result

+++ terraform/azure-servicebus-proxy/deployment_project.tf
@@ -824,8 +824,11 @@ resource "octopusdeploy_deployment_process" "deploy_project" {
         "Octopus.Action.AwsAccount.Variable" : "AWS.Account"
         "Octopus.Action.Script.ScriptBody" : <<-EOT
           OLD_STACKS=$(aws cloudformation describe-stacks --query 'Stacks[?Tags[?Key == `OctopusTransient` && Value == `True`] && Tags[?Key == `OctopusEnvironmentId` && Value == `#{Octopus.Environment.Id}`] && Tags[?Key == `OctopusProjectId` && Value == `#{Octopus.Project.Id}`] && Tags[?Key == `OctopusDeploymentId` && Value != `#{Octopus.Deployment.Id}`] && Tags[?Key == `OctopusTenantId` && Value == `#{if Octopus.Deployment.Tenant.Id}#{Octopus.Deployment.Tenant.Id}#{/if}#{unless Octopus.Deployment.Tenant.Id}untenanted#{/unless}`]].{StackName: StackName}' --output text)
-          echo "$${OLD_STACKS}"
-          echo "$${OLD_STACKS}" | xargs -n1 aws cloudformation delete-stack --stack-name $1
+
+          if [[ -n "$${OLD_STACKS}" ]]; then
+            echo "$${OLD_STACKS}"
+            echo "$${OLD_STACKS}" | xargs -n1 aws cloudformation delete-stack --stack-name $1
+          fi
         EOT
         "Octopus.Action.Script.ScriptSource" : "Inline"
         "Octopus.Action.Script.Syntax" : "Bash"
